refactor(bridge): extract client lifecycle
